software2/requirements.txt
asgiref==3.5.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
click==8.3.0
colorama==0.4.6
Django==4.1
Flask==3.1.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
outcome==1.3.0.post0
psycopg2-binary==2.9.10
pycparser==2.22
PySocks==1.7.1
pytz==2022.2.1
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2022.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
